# docs/swagger.yaml
definitions:
  MDepressionKG.Compound:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  MDepressionKG.Compoundname:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  MDepressionKG.DiseaseBac:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  MDepressionKG.DiseaseBacname:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  MDepressionKG.DiseaseResponse:
    properties:
      bac:
        $ref: '#/definitions/MDepressionKG.DiseaseBac'
      bacname:
        $ref: '#/definitions/MDepressionKG.DiseaseBacname'
      relevant_disease:
        type: string
      syndrome:
        $ref: '#/definitions/MDepressionKG.Syndrome'
      type:
        type: string
    type: object
  MDepressionKG.FoodBac:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  MDepressionKG.FoodBacname:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  MDepressionKG.FoodResponse:
    properties:
      bac:
        $ref: '#/definitions/MDepressionKG.FoodBac'
      bacname:
        $ref: '#/definitions/MDepressionKG.FoodBacname'
      compound:
        $ref: '#/definitions/MDepressionKG.Compound'
      compoundname:
        $ref: '#/definitions/MDepressionKG.Compoundname'
      food:
        type: string
      type:
        type: string
    type: object
  MDepressionKG.Syndrome:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  query.Data:
    properties:
      edges:
        items:
          $ref: '#/definitions/query.Edge'
        type: array
      nodes:
        items:
          $ref: '#/definitions/query.Node'
        type: array
      sum:
        type: integer
    type: object
  query.DataResponse:
    properties:
      bacname:
        type: string
      compoundname:
        type: string
      foodid:
        type: string
      foodname:
        type: string
      modulename:
        type: string
      mount:
        type: string
      unit:
        type: string
    type: object
  query.Edge:
    properties:
      data:
        description: empty realize
        type: object
      label:
        type: string
      source:
        type: string
      target:
        type: string
    type: object
  query.Node:
    properties:
      color:
        type: integer
      data:
        description: empty realize
        type: object
      id:
        type: string
      label:
        type: string
      node_properties:
        items:
          $ref: '#/definitions/query.NodeProperties'
        type: array
      rank:
        type: number
      size:
        type: integer
      type:
        items:
          type: string
        type: array
    type: object
  query.NodeProperties:
    properties:
      title:
        type: string
      value:
        items:
          type: string
        type: array
    type: object
  query.TextResponse:
    properties:
      bac2name:
        type: string
      bac3name:
        type: string
      bac4name:
        type: string
      bacname:
        type: string
      disname:
        type: string
      ref2:
        type: string
      ref3:
        type: string
      ref4:
        type: string
      reference:
        type: string
    type: object
info:
  contact: {}
paths:
  /graph:
    get:
      parameters:
      - description: 查询数据
        in: query
        name: search_value
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/query.Data'
      summary: 查询数据(图数据)
      tags:
      - graph
  /search/mdepression:
    get:
      parameters:
      - description: 查询类型 diseases / food
        in: query
        name: search_type
        required: true
        type: string
      - description: 查询数据参数
        in: query
        name: query
        required: true
        type: string
      - description: LIMIT <= 1000
        in: query
        name: limit
        required: true
        type: integer
      - description: NO RESTRICTIONS
        in: query
        name: offset
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/MDepressionKG.FoodResponse'
            type: array
      summary: 查询 MDepression 数据(表显示)
      tags:
      - search
  /search/microbe:
    get:
      parameters:
      - description: 查询类型 text / data
        in: query
        name: search_type
        required: true
        type: string
      - description: 查询数据参数
        in: query
        name: query
        required: true
        type: string
      - description: LIMIT <= 1000
        in: query
        name: limit
        required: true
        type: integer
      - description: NO RESTRICTIONS
        in: query
        name: offset
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/query.DataResponse'
            type: array
      summary: 查询数据(表显示)
      tags:
      - search
swagger: "2.0"